services/cms-service/scripts/requirements.txt
fastapi
uvicorn
torch
diffusers
transformers
accelerate
pillow
imageio-ffmpeg
pydantic
